chore(sql): refactor table definitions
Split up the table definitions to be easier to read and manage. Layout
is based on what jooq would generate, but I don't really want to introduce
code generation for three tables.
