+++ fiat-sql/src/main/kotlin/com/netflix/spinnaker/fiat/permissions/sql/tables/references/Tables.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2021 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.spinnaker.fiat.permissions.sql.tables.references
+
+import com.netflix.spinnaker.fiat.permissions.sql.tables.PermissionTable
+import com.netflix.spinnaker.fiat.permissions.sql.tables.ResourceTable
+import com.netflix.spinnaker.fiat.permissions.sql.tables.UserTable
+
+val USER = UserTable.USER
+
+val PERMISSION = PermissionTable.PERMISSION
+
+val RESOURCE = ResourceTable.RESOURCE
